Recompiled with namespace and make file updated
