+++ diagnostic_feedback/public/js/translations/ar/textjs.js
@@ -1,4 +1,8 @@
 
+            (function(global){
+                var DiagnosticFeedbackXBlockI18N = {
+                  init: function() {
+                    
 
 (function(globals) {
 
@@ -169,3 +173,10 @@
 
 }(this));
 
+
+                  }
+                };
+                DiagnosticFeedbackXBlockI18N.init();
+                global.DiagnosticFeedbackXBlockI18N = DiagnosticFeedbackXBlockI18N;
+            }(this));
+
