+++ diagnostic_feedback/public/js/translations/fr/textjs.js
@@ -1,4 +1,8 @@
 
+            (function(global){
+                var DiagnosticFeedbackXBlockI18N = {
+                  init: function() {
+                    
 
 (function(globals) {
 
@@ -163,3 +167,10 @@
 
 }(this));
 
+
+                  }
+                };
+                DiagnosticFeedbackXBlockI18N.init();
+                global.DiagnosticFeedbackXBlockI18N = DiagnosticFeedbackXBlockI18N;
+            }(this));
+
